Test Discord integration
Began testing discord integration in "discord_test.py." Removed "disposable.py." "meme_flexx.py" is dprecated until further notice or a mobile port is abolutely required. "meme_tk.py" is close to deprication. Plan is to merge it's features with a discord version, still continuing as a fork of Chatding.